backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.6.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.2
oracledb>=3.2.0
firebase-admin>=6.4.0
python-multipart>=0.0.6
python-dotenv>=1.0.0